Added tabindex support to dropdown component
