+++ components/select/select-option.ts
@@ -1,35 +1,56 @@
-import {
-    Component, Input, HostBinding, HostListener, EventEmitter, ViewContainerRef, ViewChild,
-} from '@angular/core';
+import {Component, Input, HostBinding, HostListener, EventEmitter, ViewContainerRef, ViewChild, Renderer, ElementRef, Output} from '@angular/core';
+import {SuiDropdownMenuItem} from '../dropdown/dropdown-menu';
+
+export type PropertyReader<T> = (obj:T) => string;
+
+export interface ISelectRenderedOption<T> {
+    value:T;
+    readLabel:PropertyReader<T>;
+    usesTemplate:boolean;
+    templateSibling:ViewContainerRef;
+}
 
 @Component({
     selector: 'sui-select-option',
     template: `
-<span #optionRenderTarget></span>
-<span *ngIf="!useTemplate">{{ readValue(value) }}</span>
+<span #templateSibling></span>
+<span *ngIf="!usesTemplate">{{ readLabel(value) }}</span>
 `
 })
-export class SuiSelectOption {
-    @HostBinding('class.item') itemClass = true;
+export class SuiSelectOption<T> extends SuiDropdownMenuItem implements ISelectRenderedOption<T> {
+    // Sets the Semantic UI classes on the host element.
+    // Doing it on the host enables use in menus etc.
+    @HostBinding('class.item')
+    private _optionClasses:boolean;
 
     @Input()
-    public value:any;
+    public value:T;
 
-    public selected:EventEmitter<any> = new EventEmitter<any>();
+    @Output()
+    public onSelected:EventEmitter<T>;
 
-    public useTemplate:boolean = false;
+    public readLabel:(obj:T) => string;
 
-    @ViewChild('optionRenderTarget', { read: ViewContainerRef })
-    public viewContainerRef:ViewContainerRef;
+    public usesTemplate:boolean;
 
-    public readValue = (value:any) => "";
+    @ViewChild('templateSibling', { read: ViewContainerRef })
+    public templateSibling:ViewContainerRef;
 
-    constructor() {}
+    constructor(renderer:Renderer, element:ElementRef) {
+        super(renderer, element);
+
+        this._optionClasses = true;
+        this.onSelected = new EventEmitter<T>();
+
+        this.readLabel = (value:T) => "";
+
+        this.usesTemplate = false;
+    }
 
     @HostListener('click', ['$event'])
-    public click(event:MouseEvent):boolean {
+    public onClick(event:MouseEvent) {
         event.stopPropagation();
-        this.selected.emit(this.value);
-        return false;
+
+        this.onSelected.emit(this.value);
     }
 }
