+++ components/select/select.ts
@@ -1,294 +1,127 @@
-import {
-    Component, Directive, ViewChild, HostBinding, ElementRef, HostListener, forwardRef,
-    TemplateRef, ViewContainerRef, AfterContentInit, QueryList, ContentChildren,
-    ViewChildren, AfterViewInit, EventEmitter,
-} from '@angular/core';
+import {Component, ViewContainerRef, ViewChild, Output, EventEmitter, ElementRef, Renderer, forwardRef, Directive} from '@angular/core';
+import {SuiSelectBase} from './select-base';
 import {NG_VALUE_ACCESSOR, ControlValueAccessor} from '@angular/forms';
-
-import {SuiDropdownMenu} from "../dropdown/dropdown-menu";
-import {SuiSelectOption} from "./select-option";
-import {Subscription} from "rxjs";
-import {Input, Output} from "@angular/core";
-import {SuiMultiSelect, SuiMultiSelectValueAccessor} from "./multi-select";
-import {SuiSelectMultiLabel} from "./multi-select-label";
-import {KeyCode} from '../util/util';
 
 @Component({
     selector: 'sui-select',
-    exportAs: 'suiSelect',
     template: `
 <i class="dropdown icon"></i>
-<input *ngIf="isSearchable" class="search" type="text" autocomplete="off" [(ngModel)]="query">
-<!-- Single-select label -->
-<div *ngIf="!selectedOption" class="default text" [class.filtered]="query">{{ placeholder }}</div>
-<div [hidden]="!selectedOption" class="text" [class.filtered]="query || !selectedOption">
-    <span #selectedOptionRenderTarget></span>
-    <span *ngIf="!optionTemplate">{{ _searchService.readValue(selectedOption) }}</span>
+<!-- Query input -->
+<input [hidden]="!isSearchable" class="search" type="text" autocomplete="off" [(ngModel)]="query" #queryInput>
+<!-- Placeholder text -->
+<div *ngIf="!selectedOption" class="default text" [class.filtered]="!!query">{{ placeholder }}</div>
+<!-- Selected item -->
+<div class="text" [class.filtered]="!!query || !selectedOption">
+    <span #optionTemplateSibling></span>
+    <span *ngIf="!optionTemplate">{{ labelGetter(selectedOption) }}</span>
 </div>
 <!-- Select dropdown menu -->
-<div class="menu" suiDropdownMenu>
+<div class="menu" suiDropdownMenu [menuAutoSelectFirst]="isSearchable">
     <ng-content></ng-content>
-    <div *ngIf="isSearchable && !results.length" class="message">No Results</div>
+    <div *ngIf="isSearchable && availableOptions.length == 0" class="message">{{ noResultsMessage }}</div>
 </div>
-`,
-    styles: [`
-:host input.search {
-    width: 12em !important;
-}
-.selected-results {
-    display: none;
-}
-`]
+`
 })
-export class SuiSelect implements AfterContentInit, AfterViewInit {
-    @ViewChild(SuiDropdownMenu)
-    private _dropdownMenu:SuiDropdownMenu;
-    private _dropdownService:any = {};
-    private _searchService:any = {};
+export class SuiSelect<T, U> extends SuiSelectBase<T, U> {
+    public selectedOption:T;
+    private _writtenOption:U;
+    
+    @ViewChild('optionTemplateSibling', { read: ViewContainerRef })
+    private _optionTemplateSibling:ViewContainerRef;
 
-    @ViewChild('selectedOptionRenderTarget', { read: ViewContainerRef })
-    private selectedOptionContainer:ViewContainerRef;
+    @Output()
+    public selectedOptionChange:EventEmitter<U>;
 
-    @ContentChildren(SuiSelectOption)
-    private renderedOptions:QueryList<SuiSelectOption>;
-    private renderedOptionsSubscriptions:Subscription[] = [];
-
-    public selectedOption:any;
-
-    @HostBinding('class.ui')
-    @HostBinding('class.selection')
-    @HostBinding('class.dropdown')
-    searchClasses = true;
-
-    @HostBinding('attr.tabindex')
-    tabIndex = 0;
-
-    @HostBinding('class.search')
-    @Input()
-    public isSearchable:boolean = false;
-
-    @Input()
-    public placeholder:string = "Select one";
-
-    @Input()
-    public get options():any[] {
-        return this._searchService.options;
+    @Output()
+    public get ngModelChange() {
+        return this.selectedOptionChange;
     }
 
-    public set options(value:any[]) {
-        this._searchService.options = value;
-        if (this.options.length > 0 && !this.options.find(o => o == this.selectedOption)) {
-            this.selectedOption = this.options.find(o => this.selectedOption == this._searchService.deepValue(o, this.keyField));
+    protected optionsUpdateHook() {
+        if (this._writtenOption && this.options.length > 0) {
+            this.selectedOption = this.options.find(o => this._writtenOption == this.valueGetter(o));
+            if (this.selectedOption) {
+                this._writtenOption = null;
+                this.drawSelectedItem();
+            }
         }
     }
 
-    @Input()
-    public get displayField() {
-        return this._searchService.optionsField;
+    protected queryUpdateHook() {
+        this.selectedOption = null;
     }
 
-    public set displayField(value:string) {
-        this._searchService.optionsField = value;
+    constructor(element:ElementRef, renderer:Renderer) {
+        super(element, renderer);
+
+        this.placeholder = "Select one";
+
+        this.selectedOptionChange = new EventEmitter<U>();
     }
 
-    @Input()
-    public keyField:string;
-
-    private get query():string {
-        return this._searchService.query;
-    }
-
-    private set query(value:string) {
-        this._searchService.updateQuery(value);
-        this.isOpen = true;
-        this.focusFirstItem();
-    }
-
-    @Output()
-    public selectedOptionChange:EventEmitter<any> = new EventEmitter<any>();
-
-    @Output()
-    public onItemSelected:EventEmitter<any> = new EventEmitter<any>();
-
-    @Input()
-    public optionTemplate: TemplateRef<any>;
-
-    @HostBinding('class.visible')
-    public get isVisible() {
-        return this._dropdownService.transition.isVisible;
-    }
-
-    @HostBinding('class.active')
-    public get isOpen():boolean {
-        return this._dropdownService.isOpen;
-    }
-
-    public set isOpen(value:boolean) {
-        this._dropdownService.isOpen = value;
-    }
-
-    @HostBinding('class.disabled')
-    @Input()
-    public get isDisabled():boolean {
-        return this._dropdownService.isDisabled;
-    }
-
-    public set isDisabled(value:boolean) {
-        this._dropdownService.isDisabled = value;
-    }
-
-    private get results():Array<any> {
-        return this._searchService.results;
-    }
-
-    private get availableOptions():Array<any> {
-        return this.results;
-    }
-
-    constructor(private el:ElementRef) {
-        this._dropdownService.dropdownElement = el;
-        this._dropdownService.autoClose = "outsideClick";
-        this._dropdownService.itemClass = "item";
-        this._dropdownService.itemSelectedClass = "selected";
-
-        this._searchService.allowEmptyQuery = true;
-        this._searchService.searchDelay = 0;
-
-        this._dropdownService.isOpenChange
-            .subscribe( (isOpen: boolean) => {
-                if (isOpen) {
-                    if (this.isSearchable && !this._dropdownService.selectedItem) {
-                        this._dropdownService.selectNextItem();
-                    }
-                }
-        });
-    }
-
-    public ngAfterContentInit():void {
-        //Initialise initial results
-        this.renderedOptionsSubscribe();
-        this.renderedOptions.changes.subscribe(() => this.renderedOptionsSubscribe());
-    }
-
-    public ngAfterViewInit():void {
-        this._dropdownMenu.service = this._dropdownService;
-    }
-
-    private renderedOptionsSubscribe() {
-        this.renderedOptionsSubscriptions.forEach((s) => s.unsubscribe());
-        this.renderedOptionsSubscriptions = [];
-
-        this.renderedOptions.forEach((option:SuiSelectOption) => {
-            this.renderedOptionsSubscriptions.push(option.selected.subscribe((value:any) => {
-                this.selectOption(value);
-            }));
-
-            setTimeout(() => {
-                option.useTemplate = !!this.optionTemplate;
-                option.readValue = v => this._searchService.readValue(v);
-
-                if (option.useTemplate) {
-                    option.viewContainerRef.clear();
-                    option.viewContainerRef.createEmbeddedView(this.optionTemplate, { option: option.value });
-                }
-            });
-        });
-    }
-
-    private renderSelectedItem() {
-        if (this.selectedOption && this.optionTemplate) {
-            this.selectedOptionContainer.clear();
-            this.selectedOptionContainer.createEmbeddedView(this.optionTemplate, { option: this.selectedOption });
-        }
-    }
-
-    public selectOption(option:any):void {
+    public selectOption(option:T) {
         this.selectedOption = option;
+        this.selectedOptionChange.emit(this.valueGetter(option));
 
-        let keyed = this._searchService.deepValue(option, this.keyField);
-        this.selectedOptionChange.emit(keyed);
-        this.onItemSelected.emit(keyed);
+        this.dropdownService.setOpenState(false);
 
-        this._searchService.updateQuery(this._searchService.readValue(option), false);
-        this._dropdownService.isOpen = false;
-        this.renderSelectedItem();
+        this.searchService.searchDelay = this._menu.menuTransitionDuration;
+        this.searchService.updateQueryDelayed("");
 
-        this._searchService.updateQuery("", false)
+        this.drawSelectedItem();
+
+        this.focusInput();
     }
 
-    private focusSearch() {
-        if (this.isSearchable) {
-            this._dropdownService.dropdownElement.nativeElement.querySelector("input").focus();
-        }
-    }
-
-    private focusFirstItem() {
-        setTimeout(() => {
-            this._dropdownService.selectedItem = null;
-            this._dropdownService.selectNextItem();
-        });
-    }
-
-    public writeValue(value:any) {
-        if (value !== null && value !== undefined) {
-            this.selectedOption = value;
+    public writeValue(value:U) {
+        if (value != null) {
             if (this.options.length > 0) {
-                this.selectedOption = this.options.find(
-                    (o:any) => value == this._searchService.deepValue(o, this.keyField));
+                this.selectedOption = this.options.find(o => value == this.valueGetter(o));
+            }
+            if (!this.selectedOption) {
+                this._writtenOption = value;
             }
         }
-        this.renderSelectedItem();
+
+        this.drawSelectedItem();
     }
 
-    @HostListener('click', ['$event'])
-    public click(event:MouseEvent):boolean {
-        event.stopPropagation();
-
-        if (!this._dropdownService.menuElement.nativeElement.contains(event.target)){
-            if (!this.isOpen) {
-                this.isOpen = true;
-                this._searchService.search();
-                this.focusSearch();
-            }
-            else if ((<Element> event.target).tagName != "INPUT") {
-                this.isOpen = false;
-            }
-        }
-        return false;
-    }
-
-    @HostListener('keypress', ['$event'])
-    public keypress(event:KeyboardEvent) {
-        if ((event.which == KeyCode.Enter || event.which == KeyCode.Space) && !this.isOpen) {
-            this.click(<any>event);
-            event.preventDefault();
+    private drawSelectedItem() {
+        if (this.selectedOption && this.optionTemplate) {
+            this.drawTemplate(this._optionTemplateSibling, this.selectedOption);
         }
     }
 }
 
-export const CUSTOM_VALUE_ACCESSOR: any = {
+// Value accessor for the select.
+export const SELECT_VALUE_ACCESSOR:any = {
     provide: NG_VALUE_ACCESSOR,
     useExisting: forwardRef(() => SuiSelectValueAccessor),
     multi: true
 };
 
+// Value accessor directive for the select to support ngModel.
 @Directive({
     selector: 'sui-select',
-    host: {'(selectedOptionChange)': 'onChange($event)'},
-    providers: [CUSTOM_VALUE_ACCESSOR]
+    host: {
+        '(selectedOptionChange)': 'onChange($event)'
+    },
+    providers: [SELECT_VALUE_ACCESSOR]
 })
-export class SuiSelectValueAccessor implements ControlValueAccessor {
+export class SuiSelectValueAccessor<T, U> implements ControlValueAccessor {
     onChange = () => {};
     onTouched = () => {};
 
-    constructor(private host:SuiSelect) {}
+    constructor(private host:SuiSelect<T, U>) {}
 
-    writeValue(value: any): void {
+    writeValue(value:U) {
         this.host.writeValue(value);
     }
 
-    registerOnChange(fn: () => void): void { this.onChange = fn; }
-    registerOnTouched(fn: () => void): void { this.onTouched = fn; }
+    registerOnChange(fn:() => void) {
+        this.onChange = fn;
+    }
+    registerOnTouched(fn:() => void) {
+        this.onTouched = fn;
+    }
 }
-
-export const SUI_SELECT_DIRECTIVES = [SuiSelect, SuiSelectOption, SuiSelectValueAccessor, SuiMultiSelect, SuiSelectMultiLabel, SuiMultiSelectValueAccessor];
